Remove active account when deleting account

+++ backend/routes/account/delete_account.py
@@ -5,7 +5,9 @@ from flask_login import login_required, current_user
 
 from accounts import Account, AccountRepository
 from routes.helpers import create_simple_response
+from users import UserRepository, User
 
+user_repository = UserRepository()
 account_repository = AccountRepository()
 
 @login_required
@@ -17,8 +19,15 @@ def delete_account(account_number) -> Response:
     if not account:
         return create_simple_response("accountNotExist", HTTPStatus.NOT_FOUND)
 
+    user: User = user_repository.find_by_id(str(account.user))
+    if not user:
+        return create_simple_response("userNotExist", HTTPStatus.NOT_FOUND)
+
     if not (str(account.user) == str(current_user.get_id())):
         return create_simple_response("unauthorisedAccountAccess", HTTPStatus.UNAUTHORIZED)
+
+    if str(user.active_account) == str(account.id):
+        user_repository.update(str(user.id), {'active_account': None})
 
     account_repository.delete(str(account.id))
     return create_simple_response("accountDeleteSuccessful", HTTPStatus.OK)
